src/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
pyotp
pillow
torch
python-multipart
aiofiles
email-validator